backend/requirements.txt
fastapi
uvicorn
python-multipart
supabase
anthropic
tribev2
torch
torchaudio
ffmpeg-python
numpy
scipy
mnemo-sdk[all]
python-dotenv
httpx
pydantic
